mqtt2strapi/requirements.txt
requests==2.25.1
paho_mqtt==1.6.1
python-dotenv==1.0.0
PyYAML==6.0
